backend/requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
requests==2.31.0
uvicorn==0.30.1
apscheduler==3.10.4
sqlalchemy==1.4.52
psycopg2==2.9.9
python-dotenv==1.0.0